Create Front-End.html
This HTML page has a form that asks the user for their start and end locations. When the form is submitted, it sends a GET request to the Flask app at the /locations endpoint with the start and end location parameters in the URL. The Flask app processes the request and returns a JSON response containing a list of locations. The HTML page then displays the list of locations in a <ul> element. If no locations are found, it displays a message saying so.

You can save this HTML page as index.html in the same directory as your Flask app, and then run the Flask app using the flask run command. When you go to http://localhost:5000 in your web browser, you should see the HTML page with the location form.