Update List-Generator.py
In this modified app, the index function renders the index.html template, which contains the form for entering the start and end locations. When the form is submitted, it sends a GET request to the /locations endpoint with the start and end location parameters. The locations function extracts the start and end locations from the request and passes them to the find_attractions function to get the list of locations. The list of locations is then returned as a JSON response.

Make sure that the index.html file is stored in the templates folder within your Flask app directory.